Project is buildable, typedef needs update nest
I got the project buildable but decided to create a new project.
So far it builds and exports js that gets executed on the screeps server.
Due to outdated typedef it errors out.
I couldnt figure out how to place the lib folder outside of main,
but could be worse.
The only thing with the old project sbt I miss is the wrapper,
but have to look into it after a while. For now the error reporting is fine.